third_party/WebKit/LayoutTests/inspector/sources/debugger-breakpoints/set-conditional-breakpoint.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>
function oneLineTestFunction() { return 0; }
</script>

<script>
function oneLineTestFunction2() { return 0; }</script>

<script>

function testFunction()
{
    var x = Math.sqrt(10);
    console.log("Done.");
    return x;
}

var test = function()
{
    var currentSourceFrame;
    InspectorTest.setQuiet(true);
    InspectorTest.runDebuggerTestSuite([
        function testSetConditionalBreakpointThatBreaks(next)
        {
            InspectorTest.showScriptSource("set-conditional-breakpoint.html", didShowScriptSource);

            function didShowScriptSource(sourceFrame)
            {
                currentSourceFrame = sourceFrame;
                InspectorTest.addResult("Script source was shown.");
                InspectorTest.waitUntilPaused(didPause);
                InspectorTest.createNewBreakpoint(currentSourceFrame, 16, "true", true)
                    .then(() => InspectorTest.waitBreakpointSidebarPane())
                    .then(() => InspectorTest.runTestFunction());
            }

            function didPause(callFrames)
            {
                InspectorTest.addResult("Script execution paused.");
                InspectorTest.captureStackTrace(callFrames);
                InspectorTest.dumpBreakpointSidebarPane();
                InspectorTest.waitBreakpointSidebarPane().then(breakpointRemoved);
                InspectorTest.removeBreakpoint(currentSourceFrame, 16);
            }

            function breakpointRemoved()
            {
                InspectorTest.resumeExecution(didResume);
            }

            function didResume()
            {
                InspectorTest.dumpBreakpointSidebarPane()
                InspectorTest.addResult("Script execution resumed.");
                next();
            }
        },

        function testSetConditionalBreakpointThatDoesNotBreak(next)
        {
            InspectorTest.showScriptSource("set-conditional-breakpoint.html", didShowScriptSource);

            function didShowScriptSource(sourceFrame)
            {
                currentSourceFrame = sourceFrame;
                InspectorTest.addResult("Script source was shown.");
                InspectorTest.createNewBreakpoint(currentSourceFrame, 16, "false", true)
                    .then(() => InspectorTest.waitBreakpointSidebarPane())
                    .then(() => InspectorTest.runTestFunction())
                    .then(testFunctionFinished);
            }

            function testFunctionFinished(callFrames)
            {
                InspectorTest.addResult("Test function finished.");
                InspectorTest.dumpBreakpointSidebarPane();
                InspectorTest.waitBreakpointSidebarPane().then(breakpointRemoved);
                InspectorTest.removeBreakpoint(currentSourceFrame, 16);
            }

            function breakpointRemoved()
            {
                InspectorTest.addResult("Breakpoints removed.");
                InspectorTest.dumpBreakpointSidebarPane();
                next();
            }
        },
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests setting breakpoints.
</p>

</body>
</html>
