third_party/WebKit/LayoutTests/inspector/sources/debugger-breakpoints/set-conditional-breakpoint-expected.txt

CONSOLE MESSAGE: line 18: Done.
CONSOLE MESSAGE: line 18: Done.
Tests setting breakpoints.


Running: testSetConditionalBreakpointThatBreaks
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (set-conditional-breakpoint.html:17)
Breakpoint sidebar pane 

set-conditional-breakpoint.html:17    var x = Math.sqrt(10);
Breakpoint sidebar pane 

Script execution resumed.

Running: testSetConditionalBreakpointThatDoesNotBreak
Script source was shown.
Set timer for test function.
Test function finished.
Breakpoint sidebar pane 

set-conditional-breakpoint.html:17    var x = Math.sqrt(10);
Breakpoints removed.
Breakpoint sidebar pane 
